requirements.txt
Flask-Bcrypt==0.7.1
Flask==1.0.3
pytest==4.6.3
pytest-cov==2.7.1
python-dotenv==0.10.3
flake8==3.7.7
gunicorn==19.9.0
bcrypt==3.1.7
marshmallow==2.19.3
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.17.0
marshmallow-enum==1.5.1
Flask-SQLAlchemy==2.4.0
SQLAlchemy-Utils==0.36.1
WTForms==2.2.1
Flask-WTF==0.14.2
Flask-Mail==0.9.1
Flask-Caching==1.7.2
Flask-Limiter==1.0.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
redis==3.3.2
SQLAlchemy==1.3.4
Flask-Migrate==2.5.2
PyJWT==1.7.1
requests==2.21.0
psycopg2-binary==2.8.3
pytz==2019.3
tzlocal==2.0.0
Flask-Admin==1.5.6